animeinrussia_parser/pyproject.toml
[tool.poetry]
name = "animeinrussia_parser"
version = "0.1.0"
description = ""
authors = ["User user <nobody@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
scrapy = "^2.7.1"
pymongo = {extras = ["srv"], version = "^4.3.3"}
python-dotenv = "^0.21.0"
sqlalchemy = "^1.4.44"
dateparser = "^1.1.4"
psycopg2 = {extras = ["c"], version = "^2.9.5"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
